web/requirements.txt
### please keep this file alphabetically sorted ###
-e git+git://github.com/jbalogh/django-debug-cache-panel@2ff3c45be962e08f02928d021c61e7008964020a#egg=cache_panel
-e git+git://github.com/ostrovok-team/django-debug-toolbar.git@0bd88fce9da1676b0be12f97b1b1fbd823fb0fcd#egg=django_debug_toolbar
-e git+git://github.com/ostrovok-team/django-import-excel.git@e2fc124e9bc5f19c9c9df0383a9d76639b958730#egg=import_excel
-e git+git://github.com/ostrovok-team/django-mediagenerator.git@dd4f3600adb96cc9213bf2450502ab5afae38eb6#egg=mediagenerator
-e git+git://github.com/ostrovok-team/suds@dd637c061cba486fe52b9fef6ba1a58ac93f5b35#egg=suds-0.3.9
-e git+git://github.com/ostrovok-team/xact.git@65e19094#egg=xact
-e git+ssh://git@git.ostrovok.ru/common@3ac1c92aa60aa99dfef5367ac420bbb6833c2419#egg=ostrovok_common
-e git+ssh://git@git.ostrovok.ru/django-ipgeobase@614b1b1a0290#egg=django-ipgeobase
-e git+ssh://git@git.ostrovok.ru/django-pagination@97acd3c005#egg=django-pagination
-e git+ssh://git@git.ostrovok.ru/django@7d98c77#egg=django
-e git+ssh://git@git.ostrovok.ru/django_modelpatch@a90951d523c1878f8a86c05ed9c2e5d082d08080#egg=django_modelpatch
-e git+ssh://git@git.ostrovok.ru/inflector@febb20a#egg=inflector
-e git+ssh://git@git.ostrovok.ru/ostrovok_gis@26bc344dccec5992c84a6eea1b51acc5344bb5f0#egg=ostrovok_gis
# M2Crypto # using ubuntu version
Markdown==2.1.1
# PIL # using ubuntu version
South==0.7.5
WebTest==1.3.4
cMarkdown==0.1.1
datadiff==1.1.4 # provide human-readable diffs of python data structures
decorator==3.3.3
django-ajax-selects==1.2.4
django-annoying==0.7.6
django-autoslug==1.5.0
django-celery==2.5.5
django-crawler==0.1
django-excel-response==1.0
django-markitup==1.0.0
django-nose==0.1.3
django-redis-cache==0.9.4
django-redisession==0.3.1
django-sentry==1.8.10
django-timedeltafield==0.6.2
django-tinymce==1.5.1b2
django-webtest==1.5.3
elementflow==0.3.0
eventlet==0.9.16
feedparser==5.1.2
geopy==0.94.2
glob2==0.3
gunicorn==0.14.3
inflect==0.2.1
jsonpickle==0.4.0
mock==0.8.0
lockfile==0.9.1
mox==0.5.3
nose==1.1.2
phonenumbers==4.8b1
# psycopg2==2.4.5 # using patched custom-build deb packcage
pyScss==1.1.3
python-daemon==1.6
python-dateutil==2.1
python-memcached==1.48
pytils==0.2.3
pytz>=2012b
redis==2.4.13
simplejson==2.4.0
threadpool==1.2.7
six==1.1.0
trans==1.4.2
xlrd==0.7.7
xlwt==0.7.4
# прежде чем вписать какую-нибудь хрень в конец файла, прочитайте его начало
django-tastypie=0.9.11
pymorphy=0.5.6